Fix health_monitor statvfs type casts and CLAUDE.md run command
Restore `as u64` casts on statvfs fields for macOS compatibility and
fix the binary run command in CLAUDE.md to use `cargo run --bin punch`.

+++ crates/punch-gorillas/src/runners/health_monitor.rs
@@ -460,8 +460,8 @@ fn get_disk_space(path: &str) -> (u64, u64) {
             unsafe {
                 let mut stat: libc::statvfs = std::mem::zeroed();
                 if libc::statvfs(c_path.as_ptr(), &mut stat) == 0 {
-                    let total = stat.f_blocks * stat.f_frsize;
-                    let available = stat.f_bavail * stat.f_frsize;
+                    let total = stat.f_blocks as u64 * stat.f_frsize as u64;
+                    let available = stat.f_bavail as u64 * stat.f_frsize as u64;
                     return (total, available);
                 }
             }
